Added hover effect on buttons
On editor and tutorial. This has to be done via an external style sheet (elm can only do inline styling). In consequence, all styling when the button is active defers to the external css.

Altered transposition menu availability - it is only blanked out if the ABC doesn't parse.  If there's no explicit key, default to C Major.