Temporary fix to FileIO
According to the documentation, I should be able to map Maybe String in elm to a returned value through the port of either a String or null.  The latter happens when the user cancels out of a file load dialog.  However, I find that null gives me a port error.

For the momemt, I use a return value which is a String and check for the empty String.