Fixed accidentals in triplets and in broken rhythm
In both AbcPerformance and MidiPerformance, accidentals were saved to state only after the entire production was translated.  So, a later note within the triplet or the same note in a broken rhythm pair would not pick up the accidental nature if it was not explicitly marked itself (i.e. if inherited from the previous occurrence).

Added a set of unit tests for both packages.