Improved error behaviour of editor controller
After any change to the ABC text, the ABC is parsed and made available both to the MIDI player and to VexTab (for the score).  If there is any parse error then neither will display.  However there are just a couple of places where valid ABC cannot be translated to a score because of an unsupported feature in VexTab.  In these cases, the MIDI player will appear but the score will not, although the VexTab error message is not displayed.